refactor(user): model 하위 entity/dto 서브폴더 분리 —
- user/model/User.java → user/model/entity/User.java
- user/model/UserStatus.java, UserRole.java, AuthUserDetails.java → entity/
- user/model/UserDto.java → user/model/dto/UserDto.java
- 영향 받은 다수 도메인의 import 경로 일괄 갱신
  (common/config, common/jwt, hq/account, store/*, warehouse/*, purchaseorder 등)
- 도메인 공통 코딩규칙 컨벤션 §3 (model/dto + model/entity 분리) 적용
