+++ src/main/java/org/example/stockitbe/common/config/LoginFilter.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.example.stockitbe.common.model.BaseResponseStatus;
-import org.example.stockitbe.user.model.UserDto;
+import org.example.stockitbe.user.model.dto.UserDto;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.AuthenticationServiceException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
