+++ src/main/java/org/example/stockitbe/common/config/LoginSuccessHandler.java
@@ -9,8 +9,8 @@ import org.example.stockitbe.common.jwt.JwtRefresh;
 import org.example.stockitbe.common.jwt.JwtRefreshRepository;
 import org.example.stockitbe.common.jwt.JwtUtil;
 import org.example.stockitbe.common.model.BaseResponse;
-import org.example.stockitbe.user.model.AuthUserDetails;
-import org.example.stockitbe.user.model.UserDto;
+import org.example.stockitbe.user.model.entity.AuthUserDetails;
+import org.example.stockitbe.user.model.dto.UserDto;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
