+++ src/main/java/org/example/stockitbe/common/jwt/JwtAuthFilter.java
@@ -8,8 +8,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.example.stockitbe.user.UserRepository;
-import org.example.stockitbe.user.model.AuthUserDetails;
-import org.example.stockitbe.user.model.User;
+import org.example.stockitbe.user.model.entity.AuthUserDetails;
+import org.example.stockitbe.user.model.entity.User;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
