+++ src/main/java/org/example/stockitbe/hq/esg/emissionquota/EmissionQuotaController.java
@@ -3,7 +3,7 @@ package org.example.stockitbe.hq.esg.emissionquota;
 import lombok.RequiredArgsConstructor;
 import org.example.stockitbe.common.model.BaseResponse;
 import org.example.stockitbe.hq.esg.emissionquota.model.EmissionQuotaDto;
-import org.example.stockitbe.user.model.AuthUserDetails;
+import org.example.stockitbe.user.model.entity.AuthUserDetails;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
